+++ Wire/Renderer/GXRenderer/WireGXIndexBuffer.h
@@ -43,7 +43,7 @@ public:
 private:
 	void CreateBuffer(Renderer* pRenderer, UInt size, Buffer::UsageType
 		usage);
-	void Copy(const IndexBuffer* pIndexBuffer, void* pBuffer, UInt offset);
+	void Copy(const IndexBuffer* pIndexBuffer, void* pBuffer, UShort offset);
 
 	TMap<UInt, PdrDisplayList*> mDisplayLists;
